+++ src/DocumentSearch.php
@@ -83,10 +83,7 @@ class DocumentSearch extends Plugin
         return Craft::$app->view->renderTemplate(
             'document-search/settings',
             [
-                'settings' => $this->getSettings(),
-                'binaries' => [
-                    'pdftotext' => is_file($this->getSettings()->pdfToTextExecutable)
-                ]
+                'settings' => $this->getSettings()
             ]
         );
     }