+++ src/models/Settings.php
@@ -19,7 +19,6 @@ use craft\base\Model;
  */
 class Settings extends Model
 {
-    public $pdfToTextExecutable = '/usr/local/bin/pdftotext';
     public $maximumDocumentSize = 1024 * 4;
     public $indexVolumes = [];
 
@@ -32,7 +31,6 @@ class Settings extends Model
     public function rules()
     {
         return [
-            ['pdfToTextExecutable', 'string'],
             [['maximumDocumentSize'], 'integer', 'min' => 1, 'message' => 'Document size should be a positive number'],
         ];
     }